Fix table truncation issue in tests
The addition of the likes tables broke all tests where the tables
in the database are being truncated. It's unclear exactly why this
is happening, as the CASCADE option I was using should prevent the
error, relating to foreigh key constraints. However, no matter what
I did with the SQL command, the error persisted.

I eventually fixed this by swapping out TRUNCATE for DELETE, and
deleting each table individually. This feels like a cludge, as from
what I can see, TRUNCATE CASCADE should work. But it doesnt, and for
the moment, the solution I have is the best I can come up with.